+++ src/booking.js
@@ -949,7 +949,7 @@ class RecrasBooking {
                 <button>${ this.languageHelper.translate('DISCOUNT_CHECK') }</button>
             </div>
         `;
-        this.findElement('.recras-contactform').insertAdjacentHTML('beforebegin', html);
+        this.findElement('.recras-datetime').insertAdjacentHTML('afterend', html);
 
         this.findElement('.recras-discounts input').addEventListener('keydown', e => {
             if (e.key === 'Enter') {
@@ -980,6 +980,20 @@ class RecrasBooking {
         });
     }
 
+    addDateTimeSelectionHtml() {
+        let today = RecrasDateHelper.datePartOnly(new Date());
+        let html = `<form class="recras-datetime">
+            <label for="recras-onlinebooking-date">
+                ${ this.languageHelper.translate('DATE') }
+            </label>
+            <input type="text" id="recras-onlinebooking-date" class="recras-onlinebooking-date" min="${ today }" disabled autocomplete="off">
+            <label for="recras-onlinebooking-time">
+                ${ this.languageHelper.translate('TIME') }
+            </label>
+            <select id="recras-onlinebooking-time" class="recras-onlinebooking-time" disabled autocomplete="off"></select>
+        </form>`;
+        this.appendHtml(html);
+    }
     showDateTimeSelection(pack) {
         let startDate = new Date();
         let endDate = new Date();
@@ -987,12 +1001,8 @@ class RecrasBooking {
 
         return this.getAvailableDays(pack.id, startDate, endDate)
             .then(() => {
-                let today = RecrasDateHelper.datePartOnly(new Date());
-                let html = `<form class="recras-datetime">`;
-                html += `<label for="recras-onlinebooking-date">${ this.languageHelper.translate('DATE') }</label><input type="text" id="recras-onlinebooking-date" class="recras-onlinebooking-date" min="${ today }" disabled autocomplete="off">`;
-                html += `<label for="recras-onlinebooking-time">${ this.languageHelper.translate('TIME') }</label><select id="recras-onlinebooking-time" class="recras-onlinebooking-time" disabled autocomplete="off"></select>`;
-                html += '</form>';
-                this.appendHtml(html);
+                this.addDateTimeSelectionHtml();
+                this.showDiscountFields();
 
                 if (this.options.getPreFilledAmounts()) {
                     this.preFillAmounts(this.options.getPreFilledAmounts());
@@ -1040,7 +1050,6 @@ class RecrasBooking {
                                 this.selectSingleTime();
                             });
                             this.maybeDisableBookButton();
-                            this.showDiscountFields();
                         },
                     }
                 );