+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## master (unreleased)
+* Show discount fields straight from the start, not after entering date
+
 ## 0.18.0 (2019-09-28)
 * Add heading to quantity form
 * Add placeholders to quantity form inputs